+++ src/CoreManifest.php
@@ -2,9 +2,9 @@
 
 namespace LaravelUi5\Core;
 
-use LaravelUi5\Core\Ui5\AbstractLaravelUi5Manifest;
+use LaravelUi5\Core\Ui5\AbstractManifest;
 
-class CoreManifest extends AbstractLaravelUi5Manifest
+class CoreManifest extends AbstractManifest
 {
 
     protected function enhanceFragment(string $module): array