+++ tests/Fixture/Hello/HelloManifest.php
@@ -2,9 +2,9 @@
 
 namespace Tests\Fixture\Hello;
 
-use LaravelUi5\Core\Ui5\AbstractLaravelUi5Manifest;
+use LaravelUi5\Core\Ui5\AbstractManifest;
 
-class HelloManifest extends AbstractLaravelUi5Manifest
+class HelloManifest extends AbstractManifest
 {
     protected function enhanceFragment(string $module): array
     {